# backend/requests/openapi.yaml
openapi: 3.1.0
info:
  title: Instant Wellness Kits API
  description: |
    REST API for order management with automatic tax calculation and geocoding.

    **Key Features:**
    - 📍 Geocoding with Mapbox (New York State only)
    - 💰 Automatic tax calculation (state, county, city, MCTD)
    - 📊 Async CSV batch import with real-time progress
    - 📡 SSE (Server-Sent Events) for live updates
    - 📈 Detailed error reporting and job history

    **Tech Stack:**
    - Spring Boot 4.0.3
    - Spring Batch 6.0
    - PostgreSQL + Redis
    - Mapbox Geocoding API
  contact:
    name: Acheron Team
    email: support@acheron.com
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Local development server
- url: https://api.acheron.com
  description: Production server
...
paths:
...
  /orders:
    get:
      tags:
      - Orders
      summary: Get all orders
      description: Retrieves a paginated list of orders sorted by total amount
      operationId: getAllOrders
      parameters:
      - name: pageable
        in: query
        description: "Pagination parameters (page, size, sort)"
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: Orders retrieved successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
    post:
      tags:
      - Orders
      summary: Create new order
      description: Creates a new order with automatic tax calculation based on coordinates
        using native GeoJSON point-in-polygon lookup. No external API calls.
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequest"
        required: true
      responses:
        "201":
          description: Order created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          description: Invalid request or location outside NY State
...
  /api/v1/map/csv-preview:
    post:
      tags:
      - Map
      summary: Preview CSV points on map (no save)
      description: Parses CSV → GeoJSON points for map preview. Does NOT persist.
        Includes point count and subtotal sum.
      operationId: previewCsvOnMap
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: "CSV file: id, longitude, latitude, timestamp, subtotal"
              required:
              - file
      responses:
        "200":
          description: OK
          content:
            application/geo+json:
              schema:
                type: object
                additionalProperties: {}
...
  /api/v1/map/orders:
    get:
      tags:
      - Map
      summary: Get orders as GeoJSON with filters + aggregate summary
      description: |-
        Returns orders as a GeoJSON FeatureCollection with rich filtering:
        - **Bounding box**: minLat/maxLat/minLon/maxLon for viewport loading
        - **Subtotal range**: minSubtotal/maxSubtotal
        - **Tax rate range**: minTaxRate/maxTaxRate (composite rate, e.g. 0.08)
        - **Date range**: from/to (ISO datetime)
        - **Limit**: max features returned (default 10000)

        Response includes `summary` with aggregate totals for the filtered set.
      operationId: getOrdersGeoJson
      parameters:
      - name: minLat
        in: query
        description: Min latitude (bounding box)
        required: false
        schema:
          type: number
      - name: maxLat
        in: query
        description: Max latitude (bounding box)
        required: false
        schema:
          type: number
      - name: minLon
        in: query
        description: Min longitude (bounding box)
        required: false
        schema:
          type: number
      - name: maxLon
        in: query
        description: Max longitude (bounding box)
        required: false
        schema:
          type: number
      - name: minSubtotal
        in: query
        description: Min subtotal filter
        required: false
        schema:
          type: number
      - name: maxSubtotal
        in: query
        description: Max subtotal filter
        required: false
        schema:
          type: number
      - name: minTaxRate
        in: query
        description: Min composite tax rate (e.g. 0.04)
        required: false
        schema:
          type: number
      - name: maxTaxRate
        in: query
        description: Max composite tax rate (e.g. 0.09)
        required: false
        schema:
          type: number
      - name: from
        in: query
        description: "From date (ISO, e.g. 2024-01-01T00:00:00)"
        required: false
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        description: "To date (ISO, e.g. 2024-12-31T23:59:59)"
        required: false
        schema:
          type: string
          format: date-time
      - name: limit
        in: query
        description: Max features to return (default 10000)
        required: false
        schema:
          type: integer
          format: int32
          default: 10000
      responses:
        "200":
          description: GeoJSON FeatureCollection with summary
          content:
            application/geo+json:
              schema:
                type: object
                additionalProperties: {}
  /api/v1/map/orders/stats:
    get:
      tags:
      - Map
      summary: Get order stats summary
      description: "Aggregate stats: total orders, revenue, tax, avg value, orders-by-county\
        \ breakdown. Supports same filters as /orders."
      operationId: getOrderStats
      parameters:
      - name: minLat
        in: query
        required: false
        schema:
          type: number
      - name: maxLat
        in: query
        required: false
        schema:
          type: number
      - name: minLon
        in: query
        required: false
        schema:
          type: number
      - name: maxLon
        in: query
        required: false
        schema:
          type: number
      - name: minSubtotal
        in: query
        required: false
        schema:
          type: number
      - name: maxSubtotal
        in: query
        required: false
        schema:
          type: number
      - name: minTaxRate
        in: query
        required: false
        schema:
          type: number
      - name: maxTaxRate
        in: query
        required: false
        schema:
          type: number
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
...
components:
  schemas:
...
    OrderRequest:
      type: object
      description: "Order details with latitude, longitude, and subtotal"
      properties:
        latitude:
          type: number
          maximum: 90.0
          minimum: -90.0
        longitude:
          type: number
          maximum: 180.0
          minimum: -180.0
        subtotal:
          type: number
      required:
      - latitude
      - longitude
      - subtotal
    OrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        latitude:
          type: number
        longitude:
          type: number
        subtotal:
          type: number
        composite_tax_rate:
          type: number
        tax_amount:
          type: number
        total_amount:
          type: number
        timestamp:
          type: string
          format: date-time
        taxBreakdown:
          $ref: "#/components/schemas/TaxBreakdown"
        jurisdictions:
          type: array
          items:
            type: string
    TaxBreakdown:
      type: object
      properties:
        state_rate:
          type: number
        county_rate:
          type: number
        city_rate:
          type: number
        special_rates:
          type: number
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...